Plugins: categorize linux.check_idt as a malware plugin
